# docs/swagger.yaml
definitions:
  entity.Node:
    properties:
      id:
        type: integer
      ports:
        items:
          $ref: '#/definitions/entity.Port'
        type: array
    type: object
  entity.Port:
    properties:
      id:
        type: string
      node_id:
        type: integer
      value:
        type: integer
    type: object
  entity.PortConnection:
    properties:
      name:
        type: string
      source:
        $ref: '#/definitions/entity.Port'
      target:
        $ref: '#/definitions/entity.Port'
    type: object
  entity.Response:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
  entity.Segment:
    properties:
      name:
        type: string
      result:
        type: integer
      sources:
        items:
          $ref: '#/definitions/entity.Port'
        type: array
      target:
        $ref: '#/definitions/entity.Port'
    type: object
info:
  contact: {}
paths:
  /health:
    get:
      description: Returns the health status of the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Health check
      tags:
      - health
  /v1/canvas/nodes:
    get:
      description: Returns all canvas nodes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Node'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get canvas nodes
      tags:
      - canvas
  /v1/canvas/nodes/{node_id}:
    get:
      description: Returns canvas node detail
      parameters:
      - description: Node ID
        format: int64
        in: path
        name: node_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Node'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get canvas node detail
      tags:
      - canvas
  /v1/canvas/ports:
    get:
      description: Returns all canvas ports
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Port'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get canvas ports
      tags:
      - canvas
  /v1/canvas/ports/{port_id}:
    get:
      description: Returns canvas port detail
      parameters:
      - description: Port ID
        in: path
        name: port_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Port'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get canvas port detail
      tags:
      - canvas
  /v1/canvas/ports/connections:
    get:
      description: Returns all canvas port connections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.PortConnection'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get canvas port connections
      tags:
      - canvas
  /v1/segments:
    get:
      description: Returns all segments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entity.Response'
            - properties:
                data:
                  properties:
                    count:
                      type: integer
                    segments:
                      items:
                        $ref: '#/definitions/entity.Segment'
                      type: array
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get segments
      tags:
      - segments
swagger: "2.0"